Extract common code from this packge into a file called helpers.go
