+++ validator/unknown_attribute.go
@@ -1,7 +1,6 @@
 package validator
 
 import (
-	"encoding/json"
 	"fmt"
 
 	microerror "github.com/giantswarm/microkit/error"
@@ -40,23 +39,6 @@ func ToUnknownAttribute(err error) UnknownAttributeError {
 	return errgo.Cause(err).(UnknownAttributeError)
 }
 
-// StructToMap is a helper method to convert an expected request data structure
-// in the correctly formatted type to UnknownAttributes.
-func StructToMap(s interface{}) (map[string]interface{}, error) {
-	b, err := json.Marshal(s)
-	if err != nil {
-		return nil, microerror.MaskAny(err)
-	}
-
-	var m map[string]interface{}
-	err = json.Unmarshal(b, &m)
-	if err != nil {
-		return nil, microerror.MaskAny(err)
-	}
-
-	return m, nil
-}
-
 // UnknownAttribute takes an arbitrary map and a map obtaining some expected
 // structure. The first argument might represent an incoming request of some
 // microservice. The second argument should then represent the datastructure of
@@ -64,10 +46,10 @@ func StructToMap(s interface{}) (map[string]interface{}, error) {
 // contains fields which are not available in expected, an UnknownAttributeError
 // is returned.
 func UnknownAttribute(received, expected map[string]interface{}) error {
-	for r, _ := range received {
+	for r := range received {
 		var found bool
 
-		for e, _ := range expected {
+		for e := range expected {
 			if e == r {
 				found = true
 				break
